requirements.txt
streamlit
pandas
numpy
matplotlib
prophet
python-dotenv
google-generativeai